refactor: revert write version
