+++ cognite/pygen/_core/logic.py
@@ -5,16 +5,12 @@ import json
 from typing import Any
 
 from cognite.client.data_classes import data_modeling as dm
-from cognite.client.data_classes.data_modeling.views import (
-    SingleHopConnectionDefinitionApply,
-    ConnectionDefinitionApply,
-)
 
 
 def _unique_properties(
-    prop: dm.MappedPropertyApply | SingleHopConnectionDefinitionApply | ConnectionDefinitionApply,
+    prop: dm.MappedProperty | dm.SingleHopConnectionDefinition | dm.ConnectionDefinition,
 ) -> dict[str, Any]:
-    if isinstance(prop, dm.MappedPropertyApply):
+    if isinstance(prop, dm.MappedProperty):
         return {
             "container": prop.container.dump(),
             "container_property_identifier": prop.container_property_identifier,
@@ -23,7 +19,7 @@ def _unique_properties(
             "nullable": prop.nullable,
             "type": prop.type.dump(),
         }
-    elif isinstance(prop, SingleHopConnectionDefinitionApply):
+    elif isinstance(prop, dm.SingleHopConnectionDefinition):
         return {
             "direction": prop.direction,
             "name": prop.name,
@@ -33,7 +29,7 @@ def _unique_properties(
         raise ValueError(f"Unknown property type {prop}")
 
 
-def _unique_views_properties(view: dm.ViewApply) -> dict[str, Any]:
+def _unique_views_properties(view: dm.View) -> dict[str, Any]:
     """
     Returns the properties from a view that uniquely defines it.
 
@@ -55,7 +51,7 @@ def _unique_views_properties(view: dm.ViewApply) -> dict[str, Any]:
     }
 
 
-def get_unique_views(*views: dm.ViewApply) -> list[dm.ViewApply]:
+def get_unique_views(*views: dm.View) -> list[dm.View]:
     view_hashes = set()
     unique_views = []
     for view in views:
