+++ tests/test_unit/test_generator/test_sdk_generator_pump_sdk.py
@@ -1,7 +1,7 @@
 import pytest
 from cognite.client import data_modeling as dm
 
-from cognite.pygen import SDKGenerator
+from cognite.pygen._core.generators import SDKGenerator
 from cognite.pygen.config import APIClassNaming, Case, Naming, Number, PygenConfig
 from cognite.pygen.exceptions import NameConflict
